Add a function to ignore uninstalling.

+++ lib/donkiUtil.rb
@@ -27,8 +27,10 @@ class DonkiUtil
 
   def removeRegisteredRepos
     @registered_repos.each do |repo|
-      _, repo_name, _, target_dir = parseRepositoryInfo(repo)
-      removeDir(insertSlash(switchTargetDir(target_dir), repo_name))
+      _, repo_name, _, target_dir, ignore_uninstall = parseRepositoryInfo(repo)
+      unless ignore_uninstall
+        removeDir(insertSlash(switchTargetDir(target_dir), repo_name))
+      end
     end
   end
   private :removeRegisteredRepos
@@ -78,16 +80,18 @@ class DonkiUtil
         repo_name = getRepoName(repo_url)
       end
 
-      repo_branch = repo['branch'] if repo.key?('branch')
-      target_dir  = repo['target'] if repo.key?('target')
+      repo_branch      = repo['branch'] if repo.key?('branch')
+      target_dir       = repo['target'] if repo.key?('target')
+      ignore_uninstall = repo['ignore_uninstall'] if repo.key?('ignore_uninstall')
     else
-      repo_url    = repo
-      repo_name   = getRepoName(repo)
-      repo_branch = nil
-      target_dir  = nil
+      repo_url         = repo
+      repo_name        = getRepoName(repo)
+      repo_branch      = nil
+      target_dir       = nil
+      ignore_uninstall = nil
     end
 
-    return repo_url, repo_name, repo_branch, target_dir
+    return repo_url, repo_name, repo_branch, target_dir, ignore_uninstall
   end
   private :parseRepositoryInfo
 end
